Updated README.md for clarity
Updated README.md for clarity, in explaining sequences.

Original Text:

"Note how many of the sequential fields start at zero (0), like how the first / ground floor is level zero in the U.K. and Europe. If a user has a wallet that displays them, the derivation path sequence for a BIP-44 compliant bitcoin wallet, in which there is a second change address in the third account, would look like this: m/44'/0'/2'/1/1."

Revised Text:

"Note that many of the sequential fields start at zero (0), like how the first / ground floor of a building is level zero in the U.K. and Europe. 

Practical Example: A user has a wallet that displays the derivation path sequence for a **BIP-44** compliant **bitcoin** wallet, and wants to locate the **second** change address in their **third** account. The **second change address** in the **third account**, would look like this: `m/44'/0'/2'/1/1`."