Update Mycelium data
The Mycelium wallet for iPhone is by far not as feature rich as the Android version and does not support SegWit but both wallets support multiple accounts, not only the /0'/0'. Right now the best "documentation" of which derivation paths are being used is the code itself. The "Docs" link was 404, so I removed it.